requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.9
python-slugify==4.0.0
pytube3==9.6.4
requests==2.24.0
selenium==3.141.0
text-unidecode==1.3
typing-extensions==3.7.4.2
urllib3==1.25.9
